requirements.txt
psycopg2-binary==2.9.5
python-dotenv==1.0.0
xlrd==2.0.1
